Moving persistence to server only
